+++ kge/config.py
@@ -783,6 +783,9 @@ def _process_deprecated_options(options: Dict[str, Any]):
                     renamed_keys.add(key)
         return renamed_keys
 
+    # 04.02.22
+    delete_key_with_value("hyperband_search.num_workers", 1)
+
     # 03.11.21
     delete_key_with_value("job.distributed.load_batch", True)
     rename_key(
@@ -795,6 +798,9 @@ def _process_deprecated_options(options: Dict[str, Any]):
         "job.distributed.combine_mirror_blocks",
         "job.distributed.stratification.combine_mirror"
     )
+
+    # 08.09.21
+    rename_key("entity_ranking.tie_handling", "entity_ranking.tie_handling.type")
 
     # 21.04.21
     delete_key_with_value("num_workers", 1)